fix lint

+++ packages/tokamak/contracts-bedrock/scripts/verification/UpgradeL1ContractVerification.s.sol
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.15;
 
-import 'forge-std/Script.sol';
-import 'forge-std/console.sol';
-import 'src/tokamak-contracts/verification/L1ContractVerification.sol';
-import '@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol';
-import '@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol';
-import {IMultiSigWallet} from '../../src/tokamak-contracts/verification/interface/IMultiSigWallet.sol';
+import "forge-std/Script.sol";
+import "forge-std/console.sol";
+import "src/tokamak-contracts/verification/L1ContractVerification.sol";
+import "@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol";
+import "@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol";
+import { IMultiSigWallet } from "../../src/tokamak-contracts/verification/interface/IMultiSigWallet.sol";
 
 /**
  * @title UpgradeL1ContractVerification
@@ -28,12 +28,12 @@ contract UpgradeL1ContractVerification is Script {
      */
     function setUp() public {
         // Load environment variables for addresses
-        _proxyAddress = vm.envAddress('L1_CONTRACT_VERIFICATION_PROXY');
-        _verificationContractProxyAdminAddress = vm.envAddress('L1_CONTRACT_VERIFICATION_PROXY_ADMIN');
-        _multisigOwner1 = vm.envUint('MULTISIG_OWNER_1');
-        _multisigOwner2 = vm.envUint('MULTISIG_OWNER_2');
-        _multisigOwner3 = vm.envUint('MULTISIG_OWNER_3');
-        _multisigWallet = vm.envAddress('MULTISIG_WALLET');
+        _proxyAddress = vm.envAddress("L1_CONTRACT_VERIFICATION_PROXY");
+        _verificationContractProxyAdminAddress = vm.envAddress("L1_CONTRACT_VERIFICATION_PROXY_ADMIN");
+        _multisigOwner1 = vm.envUint("MULTISIG_OWNER_1");
+        _multisigOwner2 = vm.envUint("MULTISIG_OWNER_2");
+        _multisigOwner3 = vm.envUint("MULTISIG_OWNER_3");
+        _multisigWallet = vm.envAddress("MULTISIG_WALLET");
     }
 
     /**
@@ -59,27 +59,21 @@ contract UpgradeL1ContractVerification is Script {
         console.log("New L1ContractVerification implementation deployed at:", address(newImplementation));
 
         // Get current implementation for comparison
-        address currentImplementation = proxyAdmin.getProxyImplementation(TransparentUpgradeableProxy(payable(_proxyAddress)));
+        address currentImplementation =
+            proxyAdmin.getProxyImplementation(TransparentUpgradeableProxy(payable(_proxyAddress)));
         console.log("Current implementation:", currentImplementation);
 
         // Prepare the upgrade transaction data
-        bytes memory upgradeData = abi.encodeWithSelector(
-            ProxyAdmin.upgrade.selector,
-            _proxyAddress,
-            address(newImplementation)
-        );
+        bytes memory upgradeData =
+            abi.encodeWithSelector(ProxyAdmin.upgrade.selector, _proxyAddress, address(newImplementation));
 
         // Get the multisig contract
         IMultiSigWallet multiSig = IMultiSigWallet(_multisigWallet);
 
         // Owner 1 submits the transaction
         vm.broadcast(_multisigOwner1);
-        multiSig.submitTransaction(
-            _verificationContractProxyAdminAddress,
-            0,
-            upgradeData
-        );
-        uint txIndex = multiSig.getTransactionCount() - 1;
+        multiSig.submitTransaction(_verificationContractProxyAdminAddress, 0, upgradeData);
+        uint256 txIndex = multiSig.getTransactionCount() - 1;
         console.log("Transaction submitted by owner 1, txIndex:", txIndex);
 
         // Owner 2 confirms the transaction
@@ -98,7 +92,8 @@ contract UpgradeL1ContractVerification is Script {
         console.log("Transaction executed by owner 1");
 
         // Verify the upgrade was successful
-        address newImplementationAddress = proxyAdmin.getProxyImplementation(TransparentUpgradeableProxy(payable(_proxyAddress)));
+        address newImplementationAddress =
+            proxyAdmin.getProxyImplementation(TransparentUpgradeableProxy(payable(_proxyAddress)));
         console.log("New implementation confirmed:", newImplementationAddress);
         require(newImplementationAddress == address(newImplementation), "Upgrade failed");
     }
